# Reuse size, header, blob decode byte buffers