On my amd64 macbook pro with go1.7
```
benchmark                       old ns/op      new ns/op      delta
BenchmarkDecode-4               2644119879     1732578144     -34.47%
BenchmarkDecodeConcurrent-4     2772070000     1807731956     -34.79%
```

go1.6
```
benchmark                       old ns/op      new ns/op      delta
BenchmarkDecode-4               3263413334     1998678477     -38.75%
BenchmarkDecodeConcurrent-4     3355724739     2145554965     -36.06%
```

benchmarks run with `go test -bench . --benchtime 10s` so the numbers are 3-5 run averages.

I don't know why this is such a large improvement. I tried a bunch of more/similar things and nothing changed the performance like this.