@paulmach 2016-08-19
Updated.  The 32megs is an upper bound so I don't have to grow the buffer ever. Could try using `bytes.Buffer` to handle a lot of that. I think the 32megs is fine since it's only once.