@paulmach 2016-08-19
I should add, to your point, in the file I tested on it was in the 2megs range.