@AlekSi 2016-08-19
[Every change breaks someone's workflow](https://xkcd.com/1172/). 😄 People may run hundreds of decoders in parallel and may not expect increased memory consumption. So yeah, let's use `bytes.Buffer.Grow()`.

Great improvement!