@paulmach 2016-08-19
Updated so header and blob buffers use bytes.Buffer. Unless I'm doing something totally wrong i'm not seeing any speed improvement anymore. 
```
BenchmarkDecode-4               1721956082     2514611990     +46.03%
BenchmarkDecodeConcurrent-4     1795521119     2687821585     +49.70%
```
This is what I was seeing in general. This one change was a huge win, other similar change had minimal if any impact. I don't know why.