@jongillham 2016-08-21
@paulmach thanks for all the work. I haven't looked in depth but are there concurrency issues with this? I think it is possible for the buffers to be called from multiple goroutines and I am not sure it is safe to do that.

Have you thought about trying [`sync.Pool`](https://golang.org/pkg/sync/#Pool)?