@paulmach 2016-08-23
Yes, this is safe, tests also pass with `go test -race`. 

`decoder.Start()` launches a [goroutine](https://github.com/qedus/osmpbf/blob/dd8b94fbc7b435865538fca3850fd41a3070d02b/decode.go#L143) that sequentially reads the fileblocks/blobs and passes them to the concurrent "decoders". The buffer used to hold the file byte data before it gets protobuf decoded into a `*OSMPBF.Blob` is what is being reused here.  Any decoded data is copied into the proto struct by the proto.Unmarshal method (I tested this and looked at code).  ie. When protobuf decodes the bytes it copies most of it into protobuf struct's `blob.ZlibData []byte` attribute. This copy is then used by the decoders, no any part of the reused buffer.

It would be interesting to use a `sync.Pool` for this `blob.ZlibData` and then for the decoded zlib data (before if gets protobuf decoded again), but when I was trying other/similar improvements they didn't make much, if any, improvement. 

Again, I don't know why preallocating a full 32 meg buffer results in such a dramatic speed improvement, but it is significant. 

Note that the speed improvement is in the [first commit](https://github.com/qedus/osmpbf/pull/15/commits/d14b73d1c237e97a216e002eb0adcb25c73e5903) in this PR. The second commit does the cleaner implementation that does not result in much of a change in performance.