@paulmach 2016-09-01
I rolled back the [byte.Buffer stuff](https://github.com/qedus/osmpbf/compare/master...paulmach:bytes-buffer) that didn't make a difference on performance. 

The pull being requested is the preallocation of the buffers that saves 33% on benchmarks.