@jongillham 2016-09-03
I don't quite like the idea of preallocating such a large 32 Mb buffer space when most uses will probably not require anywhere near that amount. So I tried:

```
func (d *Decoder) buffer(size int) []byte {
       if len(d.buf) < size {
               d.buf = make([]byte, size)
       }
       return d.buf[:size]
}
```

I started with an initial buffer size of 0 bytes. The buffer size was increased 10 times to a maximum size of 500k with the test data. However I only saw a 2 - 3% decrease in execution time. If I initialized the buffer big enough so that it was never reallocated (i.e. 500k) I also only saw maybe 10 - 14% decrease in execution time. However with the changes made in this pull request, I saw 40% decrease in execution time as described by @paulmach which is pretty spectacular! I saw similar speed having one member variable `buf` instead of `sizeBuf`, `headerBuf` and `blobBuf`.

@AlekSi what are your thoughts on preallocating a `maxBlobSize` buffer for the decoder to use? I don't see why having a gradually increasing buffer size negates most performance improvements.