@AlekSi 2016-10-23
So I took a look. And then again. And again. In the end, I exposed a method to set buffer size. See #15.

> I don't see why having a gradually increasing buffer size negates most performance improvements.

Me too. And that bothers me.