@AlekSi 2016-10-25
#19 was merged. @paulmach, thanks again.